requirements.txt
faiss-cpu
fastapi
google-genai
httpx
numpy
pydantic-settings
pymupdf
scikit-learn
sentence-transformers
InstructorEmbedding
uvicorn[standard]==0.27.1
requests==2.31.0
langchain-core
langchain-community
langchain-text-splitters
langchain-together
chromadb
langchain-chroma
unstructured
unstructured[pdf]
python-multipart
fastapi-cache2